Added 2nd recipe for trinkets
Trinkets can be repaired with a cristalis lazuli gem. The recipe doesn't check nbt, so it means you can craft a trinket with a trinket not (fully) damaged. This recipe is made so that we don't have to keep strings/copper to build a new item, and have another texture for broken trinkets which is nice. Also fixed the recipe of the xp saver that was using lapis instead of string.